make the pointers to normalized keys unsigned char * everywhere
The normalized keys code is mostly signedness-agnostic, since they are
usually accessed with memcmp() / memcpy() / etc. However, this is not
the case with the marker values, which need to be compared with
predefined constants, assigned to ints, etc. Therefore, for these to
work properly and to make it unlikely that a silly mistake will be
made down the road, we always operate on them via unsigned char
pointers.