decompression: assume compression never allows the block to expand
Add a special case to the decompression code where, if the compressed
block has the same size as the uncompressed block size, then it is
assumed to be uncompressed. In other words, assume that the
compression code will not allow the compressed data to be of the same
size or more than the uncompressed data, but will simply store the
uncompressed data in that case.

Happily enough, this allows us to remove the ifdefs for the case when
we don't use LZO, and have unified code either way.