do not attempt to decompress past what's already been compressed
Currently the decompression code crashes if it attempts to request a
compression map past the point where the extent has been compressed
(which could happen when the requested block straddles that point).
This change attempts to fix this by not requesting anything beyond
that point, and ensuring that the pages after it are already
up-to-date.